fixed imports, added main.go ,restructured auth service
