+++ services/auth/controllers/signup.go
@@ -1,24 +1,18 @@
-//package auth
+package controllers
 
-package main
+// package main
 
 import (
-	"github.com/Phase-R/Phase-R-Backend/auth/tools"
-	"gorm.io/gorm"
-
-	"github.com/joho/godotenv"
-
-	"github.com/Phase-R/Phase-R-Backend/db/database"
+	"github.com/Phase-R/Phase-R-Backend/auth/utils"
+	"github.com/Phase-R/Phase-R-Backend/auth/db"
 	"github.com/Phase-R/Phase-R-Backend/db/models"
 	"github.com/gin-gonic/gin"
-
-	"github.com/joho/godotenv"
 	"github.com/nrednav/cuid2"
 	"log"
 	"net/http"
 )
 
-var db *gorm.DB
+
 
 func CreateUser(ctx *gin.Context) {
 	const uniqueViolation = "23505"
@@ -32,7 +26,7 @@ func CreateUser(ctx *gin.Context) {
 
 	newUser.ID = id
 
-	hash, err := tools.PwdSaltAndHash(newUser.Password)
+	hash, err := utils.PwdSaltAndHash(newUser.Password)
 	if err != nil {
 		log.Fatal("could not hash password", err)
 	}
@@ -44,7 +38,7 @@ func CreateUser(ctx *gin.Context) {
 		return
 	}
 
-	res := db.Create(&newUser)
+	res := db.DB.Create(&newUser)
 	if res.Error != nil {
 		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -56,7 +50,7 @@ func CreateUser(ctx *gin.Context) {
 func FetchUser(ctx *gin.Context) {
 	id := ctx.Param("id")
 	var user models.User
-	res := db.Raw("SELECT * FROM users WHERE id = ?", id).Scan(&user)
+	res := db.DB.Raw("SELECT * FROM users WHERE id = ?", id).Scan(&user)
 	if res.Error != nil {
 		ctx.JSON(http.StatusInternalServerError, gin.H{"error": res.Error.Error()})
 		return
@@ -73,7 +67,7 @@ func UpdateUser(ctx *gin.Context) {
 		return
 	}
 
-	res := db.Model(&models.User{}).Where("id = ?", id).Updates(updatedUser)
+	res := db.DB.Model(&models.User{}).Where("id = ?", id).Updates(updatedUser)
 	if res.Error != nil {
 		ctx.JSON(http.StatusInternalServerError, gin.H{"error": res.Error.Error()})
 		return
@@ -85,33 +79,11 @@ func UpdateUser(ctx *gin.Context) {
 func DeleteUser(ctx *gin.Context) {
 	id := ctx.Param("id")
 
-	res := db.Delete(&models.User{}, id)
+	res := db.DB.Delete(&models.User{}, id)
 	if res.Error != nil {
 		ctx.JSON(http.StatusInternalServerError, gin.H{"error": res.Error.Error()})
 		return
 	}
 
 	ctx.JSON(http.StatusOK, gin.H{"message": "user deleted successfully"})
-}
-
-func Init() {
-	err := godotenv.Load(".env")
-	if err != nil {
-		log.Fatalf("Error loading .env file")
-	}
-}
-
-func main() {
-	Init()
-	r := gin.Default()
-	db = database.InitDB()
-	r.GET("/ping", func(c *gin.Context) {
-		c.JSON(200, gin.H{"message": "pong"})
-	})
-	r.POST("/user/new", CreateUser)
-	r.GET("/user/fetch", FetchUser)
-	err := r.Run(":5432")
-	if err != nil {
-		log.Fatal(err)
-	}
 }
